RHReliableDatagram: add RETRY header flag to fix deduping
This commit introduces a new header flag (RH_FLAGS_RETRY) that is
set when a message is retried. This flag is checked in combination
with the seen array to dedupe messages that have already been
received.

Previously, deduping ONLY used the seen array which stores the last
seen message ID for each node address. This is an issue in situations
where a receiving device is constantly powered and a transmitting
device periodically turns on, transmits, and shuts down. In this
scenario, the device that shuts down will always send a message
with the same ID since it's sequence number resets to zero when it
is power cycled. This issue is limited to devices that only send a
single message before shutting down. Sending more than one message
before shutting down would be an acceptable workaround for this
issue since only the last message ID is tracked and it is compared
with a simple equality check.

Note that this new header flag occupies one of the slots in the top
4 bits of the header flags that are reserved for the RadioHead lib.

Warning: deduping will be broken if some nodes in a network have this
change and others do not. This change requires that all nodes be
upgraded to have this new logic.