Merge to 1.81
This version added an explicit NRF52 platform and now looks very similar
to the Adafruit patches. There was one disagreement: Adafruit says the
NRF51 platform has hardware SPI, but the stock RadioHead library does
not. There seem to be some stock RadioHead considerations for the NRF52
hardware SPI implementation that do not consider the NRF51, so I'm
removing Adafruit's patch to make the NRF51 use hardware SPI for now.