Patch to allow for the ATSAMD51 to compile
Allows for M4 based boards such as the Adafruit Feather M4 Express and M4 Metro to compile. When compiling for other drivers, this file gets mistakenly included by GCC and the Arduino compiler. The RH_ASK in unusable with the ATSAMD51 in this condition, but I will come back with a fix that adds support and allows for users to also use the ASK library